fix slow ztunnel uninstall for post submit jobs (#48393)
* remove unsed feature flag from test setup

* debugging ztunnel cleanup times

* deploy ztunnel with shorter terminationGracePeriodSeconds

* add logic to skip istio-system ns cleanup

* lint fix

+++ pkg/test/framework/components/namespace/kube.go
@@ -233,9 +233,11 @@ func (n *kubeNamespace) createInCluster(c cluster.Cluster, cfg Config) error {
 		return err
 	}
 
-	n.addCleanup(func() error {
-		return c.Kube().CoreV1().Namespaces().Delete(context.TODO(), n.name, kube2.DeleteOptionsForeground())
-	})
+	if !cfg.SkipCleanup {
+		n.addCleanup(func() error {
+			return c.Kube().CoreV1().Namespaces().Delete(context.TODO(), n.name, kube2.DeleteOptionsForeground())
+		})
+	}
 
 	s := n.ctx.Settings()
 	if s.Image.PullSecret != "" {
